Fixes Phyrexian Totem
I thought phyrexian totem's ability should only trigger if what dealt
damage to it was a creature however that's not the case, it should
trigger only if Phyrexian Totem itself is a creature when it's dealt
damage to. Also fix it to use checkInterveningIfClause() (please review
this if it was done correctly) so it checks on resolution as well.

Also fixes to wrongly credited cards.